Made fixes to run, however will have to refactor link since his directions are set up different from renderers and the projectile Timer system doesn't currently work: no one is deloading the sprite.

+++ 3902 Project/Sprite Renderers/RendererLists/Create Methods/CreateProjectile.cs
@@ -40,13 +40,13 @@ namespace _3902_Project
             switch ((int)_direction)
             {
                 case 0: // spawn DOWN
-                    return manager.CallProjectile(projectileName, _rendDownUp.GetPosition(), _direction, timer, speed, printScale, frameRanges);
+                    return manager.CallProjectile(projectileName, _rendDownUp.GetPosition(), (int)_direction, timer, speed, printScale, frameRanges);
                 case 1: // spawn UP
-                    return manager.CallProjectile(projectileName,_rendDownUp.GetPosition(), _direction, timer, speed, printScale, frameRanges);
+                    return manager.CallProjectile(projectileName,_rendDownUp.GetPosition(), (int)_direction, timer, speed, printScale, frameRanges);
                 case 2: // spawn RIGHT
-                    return manager.CallProjectile(projectileName, _rendRightLeft.GetPosition(), _direction, timer, speed, printScale, frameRanges);
+                    return manager.CallProjectile(projectileName, _rendRightLeft.GetPosition(), (int)_direction, timer, speed, printScale, frameRanges);
                 case 3: // spawn LEFT
-                    return manager.CallProjectile(projectileName, _rendRightLeft.GetPosition(), _direction, timer, speed, printScale, frameRanges);
+                    return manager.CallProjectile(projectileName, _rendRightLeft.GetPosition(), (int)_direction, timer, speed, printScale, frameRanges);
                 default: throw new ArgumentException("Invalid direction for CreateProjectile");
             }
         }
@@ -56,13 +56,13 @@ namespace _3902_Project
             switch ((int)_direction)
             {
                 case 0: // spawn DOWN
-                    return manager.CallProjectile(projectileName, _rendDownUp.GetPosition(), _direction, timer, speed, printScale, frameRanges);
+                    return manager.CallProjectile(projectileName, _rendDownUp.GetPosition(), (int)_direction, timer, speed, printScale, frameRanges);
                 case 1: // spawn UP
-                    return manager.CallProjectile(projectileName, _rendDownUp.GetPosition(), _direction, timer, speed, printScale, frameRanges);
+                    return manager.CallProjectile(projectileName, _rendDownUp.GetPosition(), (int)_direction, timer, speed, printScale, frameRanges);
                 case 2: // spawn RIGHT
-                    return manager.CallProjectile(projectileName, _rendRight.GetPosition(), _direction, timer, speed, printScale, frameRanges);
+                    return manager.CallProjectile(projectileName, _rendRight.GetPosition(), (int)_direction, timer, speed, printScale, frameRanges);
                 case 3: // spawn LEFT
-                    return manager.CallProjectile(projectileName, _rendLeft.GetPosition(), _direction, timer, speed, printScale, frameRanges);
+                    return manager.CallProjectile(projectileName, _rendLeft.GetPosition(), (int)_direction, timer, speed, printScale, frameRanges);
                 default: throw new ArgumentException("Invalid direction for CreateProjectile");
             }
         }
@@ -72,13 +72,13 @@ namespace _3902_Project
             switch ((int)_direction)
             {
                 case 0: // spawn DOWN
-                    return manager.CallProjectile(projectileName, _rendDown.GetPosition(), _direction, timer, speed, printScale, frameRanges);
+                    return manager.CallProjectile(projectileName, _rendDown.GetPosition(), (int)_direction, timer, speed, printScale, frameRanges);
                 case 1: // spawn UP
-                    return manager.CallProjectile(projectileName, _rendUp.GetPosition(), _direction, timer, speed, printScale, frameRanges);
+                    return manager.CallProjectile(projectileName, _rendUp.GetPosition(), (int)_direction, timer, speed, printScale, frameRanges);
                 case 2: // spawn RIGHT
-                    return manager.CallProjectile(projectileName, _rendRightLeft.GetPosition(), _direction, timer, speed, printScale, frameRanges);
+                    return manager.CallProjectile(projectileName, _rendRightLeft.GetPosition(), (int)_direction, timer, speed, printScale, frameRanges);
                 case 3: // spawn LEFT
-                    return manager.CallProjectile(projectileName, _rendRightLeft.GetPosition(), _direction, timer, speed, printScale, frameRanges);
+                    return manager.CallProjectile(projectileName, _rendRightLeft.GetPosition(), (int)_direction, timer, speed, printScale, frameRanges);
                 default: throw new ArgumentException("Invalid direction for CreateProjectile");
             }
         }
@@ -88,13 +88,13 @@ namespace _3902_Project
             switch ((int)_direction)
             {
                 case 0: // spawn DOWN
-                    return manager.CallProjectile(projectileName, _rendDown.GetPosition(), _direction, timer, speed, printScale, frameRanges);
+                    return manager.CallProjectile(projectileName, _rendDown.GetPosition(), (int)_direction, timer, speed, printScale, frameRanges);
                 case 1: // spawn UP
-                    return manager.CallProjectile(projectileName, _rendUp.GetPosition(), _direction, timer, speed, printScale, frameRanges);
+                    return manager.CallProjectile(projectileName, _rendUp.GetPosition(), (int)_direction, timer, speed, printScale, frameRanges);
                 case 2: // spawn RIGHT
-                    return manager.CallProjectile(projectileName, _rendRight.GetPosition(), _direction, timer, speed, printScale, frameRanges);
+                    return manager.CallProjectile(projectileName, _rendRight.GetPosition(), (int)_direction, timer, speed, printScale, frameRanges);
                 case 3: // spawn LEFT
-                    return manager.CallProjectile(projectileName, _rendLeft.GetPosition(), _direction, timer, speed, printScale, frameRanges);
+                    return manager.CallProjectile(projectileName, _rendLeft.GetPosition(), (int)_direction, timer, speed, printScale, frameRanges);
                 default: throw new ArgumentException("Invalid direction for CreateProjectile");
             }
         }
